requirements.txt
PySide6>=6.4.0
pdf2image>=1.16.0
Pillow>=9.0.0
pytesseract>=0.3.9
pypdf>=5.0.0